git-svn-id: file:///home/gecchi/repos@2514 9df8dbd1-7660-4e89-b1e7-fa117e7ea0c7

+++ GgafDx9Core/include/jp/ggaf/dx9core/actor/GgafDx9BoardActor.h
@@ -68,9 +68,9 @@ public:
      * @param z [“x
      */
     virtual void setGeometry(int x, int y, int z) override {
-        _x = x;
-        _y = y;
-        _z = z;
+        _x = (float)x;
+        _y = (float)y;
+        _z = (float)z;
     }
 
     /**
